Rename module 'move' to 'eval'
