+++ src-tauri/benches/search.rs
@@ -1,5 +1,5 @@
+use chusst::eval::{do_move, get_best_move_recursive, SilentSearchFeedback};
 use chusst::game::Game;
-use chusst::moves::{do_move, get_best_move_recursive, SilentSearchFeedback};
 
 #[macro_use]
 extern crate bencher;
