+++ src-tauri/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod board;
+pub mod eval;
 pub mod game;
-pub mod moves;
